# contracts/openapi/agent-public.json
{
  "components": {
    "schemas": {
...
      "ConsumeQuestionLaunchRequest": {
        "additionalProperties": false,
        "properties": {
          "launchId": {
            "format": "uuid",
            "title": "Launchid",
            "type": "string"
          }
        },
        "required": [
          "launchId"
        ],
        "title": "ConsumeQuestionLaunchRequest",
        "type": "object"
      },
...
      "CreateQuestionLaunchRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 4000,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "CreateQuestionLaunchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionLaunchPayload": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 4000,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QuestionLaunchPayload",
        "type": "object"
      },
      "QuestionLaunchPayloadEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuestionLaunchPayload"
          },
          "message": {
            "default": "Question launch consumed.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "SUCCESS",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "QuestionLaunchPayloadEnvelope",
        "type": "object"
      },
      "QuestionLaunchReceipt": {
        "additionalProperties": false,
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "launchId": {
            "format": "uuid",
            "title": "Launchid",
            "type": "string"
          }
        },
        "required": [
          "launchId",
          "expiresAt"
        ],
        "title": "QuestionLaunchReceipt",
        "type": "object"
      },
      "QuestionLaunchReceiptEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuestionLaunchReceipt"
          },
          "message": {
            "default": "Question launch prepared.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "SUCCESS",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "QuestionLaunchReceiptEnvelope",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DWP Agent Runtime",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/question-launches": {
      "post": {
        "operationId": "create_question_launch_v1_question_launches_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Auth-Session-ID",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "title": "X-Dwp-Auth-Session-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Identity-Plane",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Identity-Plane"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionLaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionLaunchReceiptEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Question Launch",
        "tags": [
          "question-launches"
        ]
      }
    },
    "/v1/question-launches/consume": {
      "post": {
        "operationId": "consume_question_launch_v1_question_launches_consume_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Auth-Session-ID",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "title": "X-Dwp-Auth-Session-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Identity-Plane",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Identity-Plane"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeQuestionLaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionLaunchPayloadEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume Question Launch",
        "tags": [
          "question-launches"
        ]
      }
    },
...
  }
}